requirements.txt
torch>=2.3
torchvision>=0.18
timm>=1.0
albumentations>=1.4
opencv-python>=4.9
pandas>=2.2
numpy>=1.26
scikit-learn>=1.4
matplotlib>=3.8
pyyaml>=6.0
tqdm>=4.66
jupyter>=1.0
notebook>=7.0
pillow>=10.3
scipy>=1.12
ultralytics>=8.2
openpyxl>=3.1

